# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vidhai ERP API — multi-site mushroom farming production orchestration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all CRM contacts
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List current raw material stock levels
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryAdjustment
      tags: [inventory]
      summary: Add or adjust stock for a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustmentInput"
      responses:
        "201":
          description: Adjustment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /fleet/vehicles:
    get:
      operationId: listVehicles
      tags: [fleet]
      responses:
        "200":
          description: All vehicles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Vehicle" }
    post:
      operationId: createVehicle
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
  /fleet/vehicles/{id}:
    patch:
      operationId: updateVehicle
      tags: [fleet]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
  /fleet/fuel-logs:
    get:
      operationId: listFuelLogs
      tags: [fleet]
      responses:
        "200":
          description: Fuel logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FuelLog" }
    post:
      operationId: createFuelLog
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FuelLogInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FuelLog" }
  /fleet/maintenance-logs:
    get:
      operationId: listMaintenanceLogs
      tags: [fleet]
      responses:
        "200":
          description: Maintenance logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MaintenanceLog" }
    post:
      operationId: createMaintenanceLog
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaintenanceLogInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceLog" }
  /fleet/usage-logs:
    get:
      operationId: listVehicleUsageLogs
      tags: [fleet]
      responses:
        "200":
          description: Usage logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VehicleUsageLog" }
    post:
      operationId: createVehicleUsageLog
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleUsageLogInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleUsageLog" }

  # ── REPORTS ──────────────────────────────────────────────────────────────────
...
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventoryMovements]
      responses:
        "200":
          description: Movement list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryMovement" }
    post:
      operationId: createInventoryMovement
      tags: [inventoryMovements]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryMovementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryMovement" }

  # ── CASING SOIL TRANSACTIONS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, type, name, company, phone, email, address, notes, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [client, vendor, other] }
        name: { type: string }
        company: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }
        createdAt: { type: string }
    ContactInput:
      type: object
      required: [type, name]
      properties:
        type: { type: string, enum: [client, vendor, other] }
        name: { type: string, minLength: 1 }
        company: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }
    ContactUpdate:
      type: object
      properties:
        type: { type: string, enum: [client, vendor, other] }
        name: { type: string, minLength: 1 }
        company: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }
    InventoryItem:
      type: object
      required: [id, materialId, materialName, unit, quantityOnHand, lastUpdated]
      properties:
        id: { type: integer }
        materialId: { type: integer }
        materialName: { type: string }
        sku: { type: ["string", "null"] }
        category: { type: string }
        unit: { type: string }
        quantityOnHand: { type: number }
        locationId: { type: ["integer", "null"] }
        locationName: { type: ["string", "null"] }
        costBasis: { type: ["number", "null"] }
        buyPricePerUnit: { type: ["number", "null"] }
        sellPricePerUnit: { type: ["number", "null"] }
        imageUrl: { type: ["string", "null"] }
        qrCode: { type: ["string", "null"] }
        lastUpdated: { type: string }
    InventoryAdjustmentInput:
      type: object
      required: [materialId, quantityDelta, reason]
      properties:
        materialId: { type: integer }
        locationId: { type: ["integer", "null"] }
        quantityDelta: { type: number }
        reason: { type: string }
        notes: { type: ["string", "null"] }
    InventoryMovement:
      type: object
      required: [id, materialId, materialName, quantityKg, createdAt]
      properties:
        id: { type: integer }
        materialId: { type: integer }
        materialName: { type: string }
        fromLocationId: { type: ["integer", "null"] }
        fromLocationName: { type: ["string", "null"] }
        toLocationId: { type: ["integer", "null"] }
        toLocationName: { type: ["string", "null"] }
        quantityKg: { type: number }
        reason: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string }
    InventoryMovementInput:
      type: object
      required: [materialId, quantityKg]
      properties:
        materialId: { type: integer }
        fromLocationId: { type: ["integer", "null"] }
        toLocationId: { type: ["integer", "null"] }
        quantityKg: { type: number }
        reason: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...
    Vehicle:
      type: object
      required: [id, name, regNo, vehicleType, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        regNo: { type: string }
        homeLocationId: { type: ["integer", "null"] }
        homeLocationCode: { type: ["string", "null"] }
        homeLocationName: { type: ["string", "null"] }
        vehicleType:
          type: string
          enum: [truck, van, motorcycle, tractor, other]
        status:
          type: string
          enum: [available, in_use, maintenance, retired]
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    VehicleInput:
      type: object
      required: [name, regNo]
      properties:
        name: { type: string }
        regNo: { type: string }
        homeLocationId: { type: ["integer", "null"] }
        vehicleType: { type: string }
        notes: { type: ["string", "null"] }
    VehicleUpdate:
      type: object
      properties:
        name: { type: string }
        status: { type: string }
        vehicleType: { type: string }
        homeLocationId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    FuelLog:
      type: object
      required: [id, vehicleId, vehicleName, vehicleRegNo, fuelDate, litres, createdAt]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        vehicleName: { type: string }
        vehicleRegNo: { type: string }
        fuelDate: { type: string }
        litres: { type: number }
        costPerLitre: { type: ["number", "null"] }
        totalCost: { type: ["number", "null"] }
        odometer: { type: ["number", "null"] }
        startKm: { type: ["number", "null"] }
        endKm: { type: ["number", "null"] }
        distanceKm: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        recordedByName: { type: ["string", "null"] }
    FuelLogInput:
      type: object
      required: [vehicleId, fuelDate, litres]
      properties:
        vehicleId: { type: integer }
        fuelDate: { type: string }
        litres: { type: number }
        costPerLitre: { type: ["number", "null"] }
        odometer: { type: ["number", "null"] }
        startKm: { type: ["number", "null"] }
        endKm: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    MaintenanceLog:
      type: object
      required: [id, vehicleId, vehicleName, vehicleRegNo, serviceDate, description, createdAt]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        vehicleName: { type: string }
        vehicleRegNo: { type: string }
        serviceDate: { type: string }
        description: { type: string }
        cost: { type: ["number", "null"] }
        nextServiceDue: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        recordedByName: { type: ["string", "null"] }
    MaintenanceLogInput:
      type: object
      required: [vehicleId, serviceDate, description]
      properties:
        vehicleId: { type: integer }
        serviceDate: { type: string }
        description: { type: string }
        cost: { type: ["number", "null"] }
        nextServiceDue: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    VehicleUsageLog:
      type: object
      required: [id, vehicleId, vehicleName, vehicleRegNo, usageDate, workType, createdAt]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        vehicleName: { type: string }
        vehicleRegNo: { type: string }
        usageDate: { type: string }
        driverId: { type: ["integer", "null"] }
        driverName: { type: ["string", "null"] }
        hoursWorked: { type: ["number", "null"] }
        workType: { type: string }
        fromLocationId: { type: ["integer", "null"] }
        toLocationId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    VehicleUsageLogInput:
      type: object
      required: [vehicleId, usageDate, workType]
      properties:
        vehicleId: { type: integer }
        usageDate: { type: string }
        driverId: { type: ["integer", "null"] }
        hoursWorked: { type: ["number", "null"] }
        workType: { type: string }
        fromLocationId: { type: ["integer", "null"] }
        toLocationId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }

    # ── REPORTS ──────────────────────────────────────────────────────────────────
...